First commit of the following listed files
First commit of:
- currents.py: for now, scripts for functions that are used in modeling
currents in electrofisiology.
- linearoperators.py: in the folder timecoupling, this is for building the
matrices or linear operators used in mtf formulations coupled with
time schemes, at the moment only a semi-implicit formulation is written.
- mtf_time_coupling.py: examples of solving mtf coupled
with ODEs in time.
- oderesolutions.py: schemes for solving ODEs in time posed in the
surface of one or multiple spheres.
- righthands.py: in the folder timecoupling, right hand sides for mtf
system coupled with ODEs in time.
- solvertemplates.py: in the folder timecoupling, templates for solving
mtf systems coupled with ODEs in time.
- A couple of empty __init__.py files